% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{BinaryTree}
\alias{BinaryTree}
\title{Generating binary tree function}
\usage{
BinaryTree(n)
}
\arguments{
\item{n}{number of nodes in the tree}
}
\value{
a list in which the element at index i is a vector containing the index/indices of the 0, 1 or 2 child nodes
}
\description{
Genarate a list encoding a binary tree
}
