% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpopTree.R
\name{fpopTree}
\alias{fpopTree}
\title{Functional Pruning Optimal Partitioning for data structures in tree}
\usage{
fpopTree(vertex_data, tree, type = "mean", weights = NULL, testMode = FALSE)
}
\arguments{
\item{vertex_data}{vector of data associated to each vertex}

\item{tree}{tree structure encoded in a list}

\item{type}{a string defining the cost model to use: "mean", "variance", "poisson", "exp", "negbin"}

\item{weights}{vector of weights (positive numbers), same size as data}

\item{testMode}{boolean. False by default. Used to debug the code}
}
\value{
a gfpop object = (changepoints, states, forced, parameters, globalCost)
\describe{
\item{\code{changepoints}}{is the vector of changepoints (we give the last element of each segment)}
\item{\code{states}}{is the vector giving the state of each segment}
\item{\code{forced}}{is the vector specifying whether the constraints of the graph are active (=1) or not (=0)}
\item{\code{parameters}}{is the vector of successive parameters of each segment}
\item{\code{globalCost}}{is a number equal to the total loss: the minimal cost for the optimization problem with all penalty values excluded}
 }
}
\description{
Functional pruning optimal partitioning with data in a tree structure
}
