% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{gauss_binaryTree}
\alias{gauss_binaryTree}
\title{Generating a binary tree with changes and gaussian cost}
\usage{
gauss_binaryTree(l_tree, k, roots = NULL)
}
\arguments{
\item{l_tree}{a list encoding a binary tree}

\item{k}{the number of changes}

\item{roots}{the position of the changes (roots of the subtrees)}
}
\value{
a list assiciated to the binary tree with changes and edge means
}
\description{
Genarate a list associated to a binary tree
}
