% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{RandomTree}
\alias{RandomTree}
\title{Generating a random tree function}
\usage{
RandomTree(n, proba = rep(1/3, 3))
}
\arguments{
\item{n}{number of nodes in the tree}

\item{proba}{probability to have 1, 2 or 3 children}
}
\value{
a list in which the element at index i is a vector containing the index/indices of the 1, 2 or 3 child nodes
}
\description{
Genarate a list encoding a random tree with n nodes and 1, 2 or 3 children
}
