% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{tree_time}
\alias{tree_time}
\title{Generating the time structure of the tree}
\usage{
tree_time(l_tree)
}
\arguments{
\item{l_tree}{A tree encoded in a list describing the parent-to-child structure}
}
\value{
a list in which the position i gives a vector with couples (index, number of children to take in position i-1 in the same list)
}
\description{
Genarate the time structure of the tree from parents to children
}
